update loading of templates to fit unified templates API
